Feature: https://trello.com/c/pTe9jn5V - dao modules configuration refactoring
