+++ devicehive-rdbms-dao/src/main/java/com/devicehive/application/RdbmsPersistenceConfig.java
@@ -1,11 +1,14 @@
 package com.devicehive.application;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
 import org.springframework.boot.autoconfigure.orm.jpa.JpaProperties;
 import org.springframework.boot.orm.jpa.EntityScan;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.DependsOn;
+import org.springframework.context.annotation.*;
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
 import org.springframework.orm.jpa.JpaVendorAdapter;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
@@ -16,6 +19,11 @@ import javax.sql.DataSource;
 import java.util.Properties;
 
 @Configuration
+@PropertySource("classpath:application-persistence.properties")
+@EnableAutoConfiguration(exclude = { JacksonAutoConfiguration.class})
+@EnableTransactionManagement(proxyTargetClass = true)
+@EntityScan(basePackages = {"com.devicehive.model"})
+@EnableAspectJAutoProxy(proxyTargetClass = true)
 public class RdbmsPersistenceConfig {
 
     @Autowired
