+++ devicehive-rdbms-dao/src/main/java/com/devicehive/dao/rdbms/ConfigurationDaoRdbmsImpl.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-@Profile({"rdbms"})
 @Repository
 public class ConfigurationDaoRdbmsImpl extends RdbmsGenericDao implements ConfigurationDao {
