+++ devicehive-rdbms-dao/src/main/java/com/devicehive/dao/rdbms/DeviceClassDaoRdbmsImpl.java
@@ -20,7 +20,6 @@ import java.util.stream.Stream;
 
 import static java.util.Optional.ofNullable;
 
-@Profile({"rdbms"})
 @Repository
 public class DeviceClassDaoRdbmsImpl extends RdbmsGenericDao implements DeviceClassDao {
