+++ devicehive-rdbms-dao/src/main/resources/application-persistence.properties
@@ -1,0 +1,17 @@
+# DATASOURCE
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.datasource.url=jdbc:postgresql://127.0.0.1/devicehive
+spring.datasource.username=postgres
+spring.datasource.password=12345
+# JPA
+spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.show-sql=false
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL9Dialect
+spring.jpa.properties.hibernate.format_sql=true
+spring.data.jpa.repositories.enabled=false
+spring.jpa.properties.hibernate.id.new_generator_mappings=true
+spring.jpa.properties.hibernate.cache.region.factory_class=com.devicehive.application.hazelcast.HazelcastCustomCacheRegionFactory
+spring.jpa.properties.hibernate.cache.use_second_level_cache=true
+spring.jpa.properties.hibernate.cache.use_query_cache=true
+spring.jpa.properties.hibernate.cache.use_minimal_puts=true
+spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
