+++ devicehive-rdbms-dao/src/main/resources/application-test-configuration.properties
@@ -1,6 +1,3 @@
-spring.profiles.active=test
-spring.profiles.include=default
-
 # DATASOURCE
 spring.datasource.driver-class-name=org.hsqldb.jdbcDriver
 spring.datasource.url=jdbc:hsqldb:mem:testdb
@@ -19,41 +16,3 @@ spring.jpa.properties.hibernate.cache.use_query_cache=false
 spring.jpa.properties.hibernate.cache.use_minimal_puts=true
 spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
 spring.jpa.properties.hibernate.hbm2ddl.import_files=initial_data.sql
-
-# EMBEDDED SERVER CONFIGURATION
-server.port=${port:9000}
-server.context-path=
-
-spring.thymeleaf.prefix=classpath:/public/
-spring.thymeleaf.suffix=.html
-
-google.identity.provider.id=1
-facebook.identity.provider.id=2
-github.identity.provider.id=3
-
-# Kafka properties
-metadata.broker.list=127.0.0.1:${kafka.port:9092}
-zookeeper.connect=127.0.0.1:${zk.port:2181}
-threads.count=1
-#Kafka producer properties
-notification.serializer.class=com.devicehive.websockets.converters.DeviceNotificationConverter
-command.serializer.class=com.devicehive.websockets.converters.DeviceCommandConverter
-#Kafka consumer properties
-group.id=devicehive
-zookeeper.session.timeout.ms=1000
-zookeeper.connection.timeout.ms=20000
-zookeeper.sync.time.ms=200
-auto.commit.interval.ms=1000
-
-hazelcast.group.name=dh-cluster
-hazelcast.port=5701
-
-flyway.enabled=false
-
-riak.host=127.0.0.1
-riak.port=8087
-riak.quorum.read.option=r
-riak.quorum.read=default
-riak.quorum.write.option=w
-riak.quorum.write=default
-
