--- devicehive-riak-dao/src/main/java/com/devicehive/application/DeviceHiveApplicationConfiguration.java
@@ -1,15 +1,0 @@
-package com.devicehive.application;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceTransactionManagerAutoConfiguration;
-import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
-
-@SpringBootApplication(exclude = { JacksonAutoConfiguration.class,
-        DataSourceAutoConfiguration.class, HibernateJpaAutoConfiguration.class,
-        DataSourceTransactionManagerAutoConfiguration.class })
-public class DeviceHiveApplicationConfiguration {
-
-
-}
