+++ devicehive-riak-dao/src/main/java/com/devicehive/dao/riak/OAuthClientDaoRiakImpl.java
@@ -12,6 +12,7 @@ import com.basho.riak.client.core.query.Location;
 import com.basho.riak.client.core.query.Namespace;
 import com.basho.riak.client.core.query.functions.Function;
 import com.basho.riak.client.core.util.BinaryValue;
+import com.devicehive.application.RiakQuorum;
 import com.devicehive.configuration.Constants;
 import com.devicehive.dao.OAuthClientDao;
 import com.devicehive.dao.riak.model.RiakOAuthClient;
@@ -20,7 +21,6 @@ import com.devicehive.vo.OAuthClientVO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
 
 import java.util.*;
@@ -28,7 +28,6 @@ import java.util.concurrent.ExecutionException;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-@Profile({"riak"})
 @Repository
 public class OAuthClientDaoRiakImpl extends RiakGenericDao implements OAuthClientDao {
