+++ devicehive-riak-dao/src/main/java/com/devicehive/dao/riak/UserNetworkDaoRiakImpl.java
@@ -7,10 +7,10 @@ import com.basho.riak.client.api.commands.kv.DeleteValue;
 import com.basho.riak.client.api.commands.kv.StoreValue;
 import com.basho.riak.client.core.query.Location;
 import com.basho.riak.client.core.query.Namespace;
+import com.devicehive.application.RiakQuorum;
 import com.devicehive.exceptions.HivePersistenceLayerException;
 import com.devicehive.dao.riak.model.UserNetwork;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
 
 import java.util.HashSet;
@@ -18,7 +18,6 @@ import java.util.List;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
 
-@Profile({"riak"})
 @Repository
 public class UserNetworkDaoRiakImpl extends RiakGenericDao {
