+++ devicehive-riak-dao/src/main/resources/application-persistence.properties
@@ -1,3 +1,5 @@
+spring.profiles.include=riak
+
 # DATASOURCE
 spring.datasource.initialize=false
 spring.batch.initializer.enabled=false
@@ -8,3 +10,5 @@ riak.quorum.read.option=r
 riak.quorum.read=default
 riak.quorum.write.option=w
 riak.quorum.write=default
+
+flyway.enabled=false
