--- devicehive-riak-dao/src/test/java/com/devicehive/dao/AbstractResourceTest.java
@@ -1,8 +1,0 @@
-package com.devicehive.dao;
-
-/**
- * Dummy class to avoid compilations failures
- * todo: implement nested logic
- */
-public class AbstractResourceTest {
-}
