+++ devicehive-riak-dao/src/test/java/com/devicehive/dao/UserNetworkDaoTest.java
@@ -12,7 +12,7 @@ import java.util.Set;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
-public class UserNetworkDaoTest extends AbstractResourceTest {
+public class UserNetworkDaoTest {
 
     @Autowired(required = false)
     UserNetworkDaoRiakImpl userNetworkDao;
