+++ devicehive-services/pom.xml
@@ -114,7 +114,7 @@
         <profile>
             <id>booted-rdbms</id>
             <activation>
-                <activeByDefault>true</activeByDefault>
+                <activeByDefault>false</activeByDefault>
             </activation>
             <properties>
                 <finalName>devicehive-rdbms-${project.version}-boot</finalName>
