+++ devicehive-services/src/main/java/com/devicehive/application/DeviceHiveApplication.java
@@ -3,21 +3,22 @@ package com.devicehive.application;
 import io.swagger.jaxrs.config.BeanConfig;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.context.web.SpringBootServletInitializer;
-import org.springframework.boot.test.SpringApplicationConfiguration;
-import org.springframework.context.annotation.*;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.FilterType;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
-import org.springframework.stereotype.Component;
 import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
 
 import javax.validation.Validator;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
-@Component
-@SpringApplicationConfiguration( classes = {DeviceHiveApplicationConfiguration.class})
+@SpringBootApplication
 @ComponentScan(value = "com.devicehive", excludeFilters = {@ComponentScan.Filter(type = FilterType.REGEX, pattern = "org.springframework.transaction.*")})
 @EnableScheduling
 @EnableAsync(proxyTargetClass = true)
@@ -26,7 +27,7 @@ public class DeviceHiveApplication extends SpringBootServletInitializer {
     public static final String MESSAGE_EXECUTOR = "DeviceHiveMessageService";
 
     public static void main(String ... args) {
-        SpringApplication.run(DeviceHiveApplication.class);
+        SpringApplication.run(DeviceHiveApplication.class, args);
     }
 
     @Override
