+++ devicehive-services/src/main/resources/application.properties
@@ -1,23 +1,4 @@
 spring.profiles.active=default
-spring.profiles.include=rdbms
-
-# DATASOURCE
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://127.0.0.1/devicehive
-spring.datasource.username=postgres
-spring.datasource.password=12345
-# JPA
-spring.jpa.hibernate.ddl-auto=validate
-spring.jpa.show-sql=false
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL9Dialect
-spring.jpa.properties.hibernate.format_sql=true
-spring.data.jpa.repositories.enabled=false
-spring.jpa.properties.hibernate.id.new_generator_mappings=true
-spring.jpa.properties.hibernate.cache.region.factory_class=com.devicehive.application.hazelcast.HazelcastCustomCacheRegionFactory
-spring.jpa.properties.hibernate.cache.use_second_level_cache=true
-spring.jpa.properties.hibernate.cache.use_query_cache=true
-spring.jpa.properties.hibernate.cache.use_minimal_puts=true
-spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
 
 # HTTP message conversion
 spring.http.converters.preferred-json-mapper=gson
@@ -63,10 +44,3 @@ management.security.sessions=STATELESS
 
 hazelcast.group.name=dh-cluster
 hazelcast.port=5701
-
-riak.host=127.0.0.1
-riak.port=8087
-riak.quorum.read.option=r
-riak.quorum.read=default
-riak.quorum.write.option=w
-riak.quorum.write=default
