+++ devicehive-services/src/test/resources/application-test.properties
@@ -1,33 +1,23 @@
-spring.profiles.active=default
-spring.profiles.include=riak
+spring.profiles.active=test
+spring.profiles.include=default
 
-# DATASOURCE
-spring.datasource.initialize=false
-spring.batch.initializer.enabled=false
+# EMBEDDED SERVER CONFIGURATION
+server.port=${port:9000}
+server.context-path=
 
 # HTTP message conversion
 spring.http.converters.preferred-json-mapper=gson
 
-server.context-path=/dh
-server.undertow.buffer-size=10240
-
 spring.thymeleaf.prefix=classpath:/public/
 spring.thymeleaf.suffix=.html
-spring.thymeleaf.mode=LEGACYHTML5
-
-# Custom configuration properties
-app.executor.size=20
-
-build.version=${project.version}
-build.timestamp=${build.timestamp}
 
 google.identity.provider.id=1
 facebook.identity.provider.id=2
 github.identity.provider.id=3
 
 # Kafka properties
-bootstrap.servers=127.0.0.1:9092
-zookeeper.connect=127.0.0.1:2181
+bootstrap.servers=127.0.0.1:${kafka.port:9092}
+zookeeper.connect=127.0.0.1:${zk.port:2181}
 command.partitions.count=3
 command.update.partitions.count=3
 device.partitions.count=3
@@ -42,15 +32,10 @@ auto.commit.interval.ms=1000
 notification.expire.sec=600
 command.expire.sec=600
 
-management.context-path=/app
-management.security.enabled=true
-management.security.role=ADMIN
-management.security.sessions=STATELESS
-
 hazelcast.group.name=dh-cluster
 hazelcast.port=5701
 
-riak.host=127.0.0.1
-riak.port=8087
-
 flyway.enabled=false
+
+
+
